Menu's, Widgets, Categories optiontype
* Added menu item category (Rolf)
* Isolated menu logic in generic location (menuitemgeneric)
* Improved menu item deletion using keyboard shortcut
* Improved icons of menu items (post/page, external link, category)
* Improved way how to add new menu items by first showing icons of
available menu items and afterwards adding new row with that specific
type, and opening properties of that type afterwards, rather than always
starting with a reference to post/page.
* Making it less easy to change types of existing menu items (buttons
are wrapped, and wrap is closed by default, making the UX more logical)
* Categories option type can now be set to a mode that only allows one
checkbox to be checked at a specific moment in time.