Helper, Page Decorator, Maps widget
* Helper; added helper functions to quickle get all posts in the DB, and
to query all widgets of a certain type for all posts
* Page Decorator; integrated self-help; if no page decorator applies for
a page, the page decorator icon will show a popup when its clicked upon.
That popup will show a link pointing to the support video on how to set
it.
* Maps widget; the "Address" label now is labeled "Locate Address", to
better show that this text box is used to locate addresses (rather than
representing the address that is displayed).