Cleanup, Menu, HtmlCustom, Taxonomy extended properties, Business site, Entities widget
* Cleanup; removed "nxs_cleanimg"; it wasn't used anymore
* Menu; "taxonomies" now renders as one item, without child elements.
Once you click on it, it will display the individual taxonomies (one
click more, but since the # of taxonomies can grow rapidly, this is the
best way forward)
* HtmlCustom widget; now supports the "code" render-behaviour
* Taxonomy extended properties; each taxonomy can now define which
extended properties it supports; those will be rendered in the backend
(for example testimonials should know the # of stars, where-as services
don't).
* Business site; added a function to retrieve instances of a particular
taxonomy; this function should be used, instead of the "getmodel"
returning all entities. This should enable the system to be initialized
faster (since not all pages need all entities to be retrieved)
* Business site; the taxonomies data can be extended by plugins,
enabling plugins to inject custom taxonomies, next to the ones injected
by the framework
* Entities widget; added support for "HmtlCustom" child widgets
* Entities widget; now allows plugins to extend the child args of the
widget to be render prior to the actual rendering (filter:
nxs_f_entity_getchildargs)