List item, Debug, Entities, Plugins, Taxonomies, Services, New Article, Menu Item Entities, ModelEditor
* List item; list style position inside text widget is now outside
instead of inside
* Debug; framework can show wpversion
* Entities; taxonomies now support show_ui parameter
* Plugins; plugins are loaded only when the site settings are set. If
site settings are not set, the plugins are ignored (preventing themes
from loading properly)
* Taxonomies; introduced the "taxonomies" taxonomy (for storing
properties on singleton abstract instances of the taxonomies like
"Services").
* Taxonomies; the abstract instance of the taxonomies are now auto
created if they don't exist.
* Services; instances now have ability to configure iconpicker
* New Article; added dependency check; only works when using WP 4.4 or
greater. If used on a version below, the function will throw an error
(only from 4.4 on the meta_fields are supported; on which the plugin of
the businesssite relies; otherwise the meta fields for globalid will not
be set, causing endless loops and endless posts to be created).
* Menu item entities; the title rendered is no longer rendeed based upon
the properties of the widget, but instead in the abstract instance of
the taxonomy itself.
* ModelEditor; the metabox now only renders the iconpicker when the
entity supports it
* Entities widget; re-ordering items is now supported directly through
the front-end.