Styling, Businessmodel, Custom CSS, SEO, Text Widget, BugLogo, Entities
* Styling; fix css bug; margin-bottom on icon right was missing
* Businessmodel; persisttype was introduced on taxonomies and taxonomy
instances in the taxonomiesmeta
* Businessmodel; the back end editing screens now use the persisttype to
render properties; some properties (like wp_title are rendered using the
default WP backend)
* Businessmodel; now implements a master-slave feature; the slave is
able to import content from a model outside the theme
* Businessmodel; implements virtual posts from the model hosted outside
the theme
* Custom css; is disabled (workaround; use customcss=true parameter); we
want people to use a custom css plugin instead of making the framework
have this feature (since we want to encourage people to NOT use custom
css as it results in support questions where people screw up their site)
* SEO; the SEO tab is no longer present in the front-end; if people want
to configure and optimize their SEO; they should use the WP backend for
that. Reason is that the SEO tab stopped working in the front-end
because some modification made outside our logic. The more we
encapsulate features like this, the more load we add to our company
which makes us less ExO, so therefore the feature is removed.
* Text Widget; can now render media items with "dynamic" dimensions when
invoked through code (entities widget). The image rendered can therefore
now be an image from an external place (using an image provider and an
image id), in an API kind of way, and using a CDN by default, and
optimal output size.
* BugLogo; now uses the improved model fields to render the company name
and slogan
* Entities widget; now support media configuration for rendering media
items from image provider(s).