Template mapping
* Template mapping;  re-implemented the modelmapping; in the old style
you could only reference a model in the business rules and the template
map to a static property of that model. In the new approach the business
rule has both a modeluris field, as well as a key/value set that
determine the actual values (@@template.xyz) of a specific property.
Benefit is that the template can now use any value, and the business
rule map each field to a specific model of choice (also supporting
static fields), also less lookups are needed (configuration becomes
simpler).