Commenting, TemplateProperties, ModelURI, List
* Commenting; lookup table definition and model definition now support
commenting by prefixing the line with // (that line wont be interpreted)
* TemplateProperties; the lookup recursion was not done properly; only
the first item would be replaced
* ModelURI; "singleton@listof" is now interpreted as an incorrect
modelURI to prevent useless roundtrips to the server
* List; the List widget now renders 4 dummy items if the user is logged
in as an admin -and- if the iterator datasource is blank, and/or if the
datasource contains a placeholder (/is invalid/). This helps in
visualizing something useful when the user is creating a template that
uses the List widget.
* List; fixed bug in the recursive evaluation of lookup values to the
items that caused lookup values to NOT be replaced
* List; now supports comments for both the lookup table and model
references for its items.
* List; the number of columns is now determined -after- all items are
rendered (its not possible to do the determination beforehand, as there
could be conditions that cause items to be filtered out, resulting in
incorrect evaluation of the number of columns in the old implementation)