Titles, RemotePosts, Shortcodes, Carousel widget, List widget
* Titles; (Kacem); each time a private or protected page "foobar" was
saved using the popup, the title would get an annoying prefix "private:
foobar", and the next time, the same; "private: private: foobar"). Fixed
by overriding the private_title_format and the protected_title_format
filters of WP.
* RemotePosts; caching increased to 30 days instead of 60 secs. Refresh
/wipe/ the cache using the modeltracker plugin
* Shortcodes; implemented a way to restrict the shortcodes from applying
"always"; in some cases we want shortcodes to apply only at a later
moment in the life cycle of creating the page (for example in the List
widget). By specifying a sc_scope attribute the shortcode can determine
whether it will, or will not be processed.
* Carousel widget; it appeared the images of the carousel didn't yet use
the theme version rewrite; causing the carousel images to not be updates
when the photopack was installed on some hosting providers; resolved.
* List widget; implemented a "where" filter allowing users to exclude
certain items from the iterator at runtime. The List widget now also
activates the "list.iterator.filter" while evaluating the shortcode
values of the lookup items while iterating over the datasource items.