Helper, ModelPicker, Select Optiontype, Embed widget, List widget
* Helper; added a function to parse keyvalues from a string of =
seperated key values in \r\n seperated lines
* ModelPicker; rewritten the modelpicker optiontype; now renders all
possible items according to the remote datasource, plus a way to use a
lookup variable. Items are sorted from A-Z (the lookup variable is
appended at the bottom)
* Select optiontype; system will now automatically append the lookup
placeholder option to allow widgets using this option to be bound to a
lookup key
* Embed widget; now has a 2 staged configuration; the first stage asks
users which "embeddable" type they want to use, and the second stage is
the stage where the actual configuration of that type occurs (based upon
parameters specified by the remote widget). The embed widget now also
supports a lookup table to allow specifying additional lookup key/values
(this allows having to create page template rules on the content server
which will slow things down and make the configuration a mess in the
long run). The embedded result/response/ is now cached for 30 days to
speed things up.
* List widget; the where clause is now a textarea as the single line was
way too small for complex filters.