Shortcodes, BusinessSite
* Shortcodes; added ucfirstchar shortcode which makes the first
character of a string uppercase
* Shortcodes; added youtubeify shortcode which converts a youtube link
in a string into an embedded youtube video
* Shortcodes; added htmlentities shortcode which converts all special
characters of the input into a htmlentities version (used to change ' to
&quot; and " to html form, see php doc on the details)
* Shortcodes; add linkify shortcode, which converts links in a string
into clickable links to that destination
* BusinessSite; greedy support; when grabbing one item from the content
server, the system will now check a greedyfetchsupport flag in the
response. If its set, the system will now automatically perform a bulk
load after the single item has been fetched. This greatly boosts the
performance of the websites when loading new content for the first time.
It can be optimized further, as it currently requires 2 cycles (one for
the first item and then a second one for all the bulk items), but well
optimized compared to how it used to work.
* BusinessSite; now "not found" models are also cached as transients (in
the old implementation not found items would not be cached, causing
useless delays in all subsequent requests that kept on repeating it).