Cache, Lookups (Shortcodes), Webmethods, PageTemplateRules, Transients, Shortcodes, HtmlCustom, BusinessSite, Caching, Embed, List
* Cache; added comments to what each section does
* Lookups (Shortcodes); the moment a lookup entry has a replaced value,
we apply the shortcodes instantly as the next step (required for rel
links)
* Webmethods; added a "raw" return function (needed to output lists for
which the OK entry is not wanted as it ruins the output)
* PageTemplateRules; the rules now allow user to again select a front
end template (it wrongly shows a list of posts and pages which was
wrong)
* Transients; added a function to wipe transients starting with a
particular prefix
* Shortcodes; no support queryparameter shortcodes (next to the now
obsolete usequeryparameter)
* Shortcodes; the queryparameter shortcode now fires a
nxs_a_usesqueryparameter to which plugins can hook
* Shortcodes; added support for the nxsstring listmodeluris shortcode; a
practical solution to allow the modelmanager to dynamically derive
subsets (indexes) of an entire model during the "auto pilot" staging.
* Shortcode; added support for the nxsstring filegetcontents shortcode;
a practical solution to output for example json API's in a list
* Shortcode; added support for the nxsstring jsonsubvalues shortcode;
used to convert a json output into a subset of itself
* Shortcode; added support for the nxsstring ifthenelse shortcode; used
to convert an input value to either a "then" attribute or a "else"
attribute based upon the evaluation of the condition provided.
* Shortcode; added support for the nxsbool httpok shortcode; used to
evaluate whether a particular http url results in an error ("false") or
whether it returns a valid response ("true").
* Shortcode; added support for the nxsbool not shortcode; it performance
a boolean "not" operator on the input (return "false" when the input is
"true", and the otherway around).
* HtmlCustom widget; now supports lookup logic
* BusinessSite; added logic to verify the validity of a schema
* Caching; the bulk (prefetch) now will prefetch both the specified
singularschema, as well as its corresponding list
* Embed widget; can now also be placed in containers other than the main
content.
* List widget; now supports a index constrained for the specified
datasource (min and max); all uri's outside the specified index range
will be ignored. This is to speed up testing particular datasets
(practical for processing large sets of data using QA reporting)
* List widget; now supports a page size (it will render a max of
specified items for which the "where" clause is true. If more items
would meet the condition those will be ignored
* List widget; now supports a lookup declaration statement that is
evaluated each time the widget renders (just 1x). The variables declared
can be used in the start template, the end template, and the iterating
template, as well as the item lookups
* List widget; now supports a begin and end html template (rendered
outside the iterating items)