UrlPrettyfy; Shortcodes, HtmlCustom, Text, List
* UrlPrettyfy; improved support for faulty inputs (////)
* Shortcodes; added the nxsstring urlfraction shortcode, used to strip
out unwanted characters for url fractions (includes stripping out /'s)
* Shortcodes; added the nxsstring modeldump shortcode used to debug
models; it outputs a json string of the content of the properties
taxonomy of the  modeluri specified
* Shortcodes; extended the nxsstring listmodeluris shortcode; can now
also list modeluris based upon the equals operator
* HtmlCustom; the lookups are now wrapped in a container, which is
closed by default
* Text widget; added a lookups input option (closed by default)
* Text widget; removed the modeluris property (in favour of the lookups
input)
* Text widget; re-implemented the applying of lookups (recursive
support, support for template rules, applied on both the text and title
props)
* List widget; added the option to not render the grid blocks; i can
imagine this will become the default behaviour instead of the exception
which it is now. Practical if the list widget should render for example
list items (<li>'s); in which case adding the div's of the flex box is
completely inpractical as it messed up the html.