UrlPrettyfy, Template Rules lookups, Shortcodes, Urls, UrlPrettyfy, Callout, GoogleMap, Quote, Seo, Text, Models, Entities, Embed
* UrlPrettyfy; now exposes an error action if it contains a variable
when being invoked
* Template Rules lookups; values are now trimmed (avoiding errors in
canonical urls)
* Shortcodes; added a function to reconstruct a shortcode based upon its
input parameters (needed on multiple occasions, therefore now a
function)
* Shortcodes; the nxsstring urlprettyfy shortcode will not do anything
with its contents if its contents still contains a {{ or [.
* Urls; nxs_geturlcurrentpage now returns the url the user was on when a
web method is invoked
* Urls; when a site runs on https (443), the url will return the pretty
url (without the port indication).
* UrlPrettyfy; now exposes the nxs_a_prettyfyurlerror action if the url
could not properly be evaluated
* Callout widget; replaced modeluris with lookup block
* Google maps; implemented support for the lookup fields
* Google maps; implemented renderstyle v2 (less clutter)
* Quote widget; implemented support for lookups section
* Seo widget; now will render in red if the current url does not equal
the value according to the evaluation of the widget.
* Text widget; we now also apply lookups translations for the
destination_url field
* Models; the ModelProperty shortcode and iterator properties now return
the htmlentities encoded value stored in the model, instead of the raw
version as it appears the shortcodes don't evaluate properly in the
lookups evaluations if a text contain special characters l(ike  for
example the less than (<) character.
* Models; if a model lookups fails the system will now cache the error
to avoid useless repeting loads to the contentprovider server
* Models; now fires "nxs_a_invalidschema" and "nxs_a_modelerror"
actions.
* Entities; entities widget is phased out in favor of the List widget.
* Embed widget; renamed to "Section"
* List widget; renamed to "Repeater"