Remote posts, Shortcodes, BusinessRuleUrl, Callout, SEO, Entities, Embed, List
* Remote posts; changed the key for the transients, such that clearing
the cache can now be performed on all items in one go when needed
* Shortcodes; added the nxsstring modelidbymd5 shortcode which can pick
a semi random id of a referenced schema by upon a indexer property
* BusinessRuleUrl; fixed bug when using a template that requires a model
for a specific schema (would crash the handling of the page)
* Callout widget; now applies lookups to the image_src property too
* SEO widget; now is taken into consideration for all requests (in the
old situation this would only be applied for virtual posts)
* Entities widget; temporarily re-enabled this as a conceptual widget so
we can better see how the sites used to work that we need to
upgrade/fix.
* Embed widget; renamed the widget back from 'Section" to "Embed", as
Section was already in use.
* Embed widget; now adds an additional "nxs_triggeredby" queryparameter
with value "Embedwdiget" to the logic on the contentprovider can take
that into consideration (used for the 301 template)
* Embed widget; the lookups were not applied recursively onto itself.
* Embed widget; the lookups were re-calculated for each field of the
widget which was useless
* List widget; now supports a json datasource too if you prefix the
iterator datasource with "json:" (used to play around with the slider
widget as an embeddable widget; the slides need to be specified by the
client in that case; that one does not necessarily have to be an
existing model).