Shortcodes, Lookups, Text widget, List widget
* Shortcodes; no longer the nxsstring modelproperty will not apply the
shortcode if the modeluri or property contains an unreplaced variable
(in the old situation it would, which would always result in a model
error action, even though it was every likely that the lookup at a later
moment in time would have evaluated just fine).
* Shortcodes; added the nxsstring html_entity_decode shortcode
* Shortcodes; the nxsstring modelidbymd5 now implements a way to
optionally
* Lookups; blending lookups to one another will NO LONGER apply
shortcodes; shortcodes should be processed /evaluated/ when the lines
are processed sequentially. If the shortcodes would be applied this can
easily result in model lookups resulting in errors as the variables are
not yet defined.
* Text widget; the lookup logic is applied on the image_src too
* List widget; changed the order at which lookup items of the iterated
item are added; they are now added prior to the item lookups.