Textarea, Shortcodes
* Textarea; textareas in the admin popups will now automatically expand
such that the entire content is visible (instead of having this tiny
window and poor scrolling required to see the lookups / content)
* Shortcodes; improved warnings for authenticated users in the
modelproperty shortcode when the user has put 2 equal signes next to one
another when assigning lookup values.
* Shortcodes; the nxsstirng modelidbymd5 shortcode will now crash when
the source schema is not found. It will also crash if it contains 0
elements.