Shortcodes, FormBox, BusinessSite, List
* Shortcodes; implemented the nxsstring min operator which returns the
lower number of a ";" delimited string. For example 2;8;1 would return
1.
* Shortcodes; implemented the nxsstring getlatlng operator which returns
the lat;lng;found values of an address; practical for auto piloting
content to the centralized storage
* Shortcodes; implemented the nxsstring intval strip functionality which
can filter allexceptdigits chars
* Shortcodes; the nxsstring modelproperty operator now also replaced
dollar signs with the html entitie &dollar;
* Shortcodes; implemented the "currentlurl" nxsstring operator
* Shortcodes; extended the "explode" nxsstring operator; can now also
return a value at a specific index
* Shortcodes; implements the "modelexists" boolean operator
* FormBox; moved the debug and write-to-model properties to the bottom
and collapsed the containers
* BusinessSite; implemented a way to prevent hammering the server; if a
certain treshhold number of models fail to be retrieved for a particular
schema, any subsequent attempts will automatically fail too (will not
even be attempted), to reduce the number of hammering on the server.
* List; pagesize can now also be filled using lookup value
* List; simplified the modeluris implementation