CSS, Page template props, Shortcodes, YouTube, Bulkmodels
* CSS; native menu parent items now also render as clickable items
* Page template props; if postid equals -1 the system wont try to
interpret it as a remote model (needed for the portfolio plugin to work)
* Shortcodes; added the nxsstring homeurl operator, returning the home
url of the site
* Shortcodes; first attempt on implementation of "smartlinks" operator
for the nxsstring  shortcode; will replace words in a value based upon a
dictionary (eventually to be drive by a model)
* Shortcodes; nxsstring listmodeluris shortcode now supports segmented
schemas
* Shortcodes; listmodeluris shortcode now supports multiple conditions
that should be met at the same time (where_property, where_property_1
and where_property_2). When needed can easily be extended for more
conditions
* Shortcodes; implemented the "isempty" nxsbool shortcode
* Shortcodes; implemented the "in_array" nxsbool shortcode
* YouTube; widget now supports lookups too
* YouTube widget; now shows the videoid and allows it to be editable (to
use lookups)
* Bulkmodels; the content server (IP 52.21.12.12) is now able to
instruct sites to bulk refresh particular schema's, and to clear the
cache of the site; can optionally be used to update sites when using the
auto pilot function when central data on the content server is
refreshed.