text area
* text area; height of text areas in popups now are recalculated, untill
a certain max value (500px) is reached. Once its reached this height, we
will no longer make it larger to prevent very annoying wobbling of the
screen if the cursor is set near the bottom of the textarea.
Recalculating the height would then cause the browser to scroll to the
top of the textarea when the height is re-set, making it practically
impossible to use the textarea.