Frontendframework, Embed, Wpmenu, BusUrl, Nxs2
* Frontendframework; body tag now get a class from the front end
framework used to render that page
* Frontendframework; customfrontendframework parameter of the lookup
tables now selects the front-end editor that is used to render the page
(can be overruled by specifying the frontendframework parameter). If not
specified it will still use the "nxs" framework by default.
* Embed; embed widget now properly aligns columns in rows, and also
renders the google maps widget properly
*  Wpmenu; clicking now renders animation in nxs2 framework
* Wpmenu; clicking on the expand/collapse button no longer moves cursor
to the top
* BusUrl; if a busurl matches, this will override the "is_404" property
of the wp_query (is_404() will no longer return true)
*  Nxs2; inline styles are now rendered above the embedded js and css
* Frontendframework; changes the way the frontendframeworks initialize;
in previous code the frontendframework was loaded prior to the init, and
during the init action it initialized. This is now postponed; the
frontendframework now initializes AFTER the init event, it initializes
as soon as the page template rules have loaded. Also at that time it
initializes the framework. This is postponed as its important that the
page template rules can set/specify which frontendframework is used to
render a page (for example some pages have to render in one
frontendframework kind of way, where-as other need other ones).