Header (WC), Footer (WC)
Header (WC); the WooCommerce triggers the get_header in the archive
pages. These caused the re-rendering of widgets to throw an error, as
the header of the page would be re-rendered (Halted; expected
nxs_global_row_render_statebag to be null, but it isn't?). Fixed.
Footer (WC); the WooCommerce triggers the get_footer in the archive
pages. These caused the re-rendering of widgets to throw an error, as
the header of the page would be re-rendered (Halted; expected
nxs_global_row_render_statebag to be null, but it isn't?). Fixed.