Shortcodes
* Shortcodes; urlprettyfy operator of the nxsstring shortcode can now
optionally also render links for a custom homeurl (other than the
homeurl of the current site). Defaults to the homeurl of the current
site.
* Shortcodes; added the modelurisbymd5 operator for the nxsstring
shortcode; can list a set of semi random modeluris from a particular
schema (amount is determined by min and max parameters), supporting
cachebehaviour for the schema. The resulting uris will be distinct.