Popups, Debug, Page popup widget, Article Link optiontypes, Formbox
* Popups; after the page popup was rendered, all sequential popups were
stretched (not properly aligned). Fixed.
* Popups; implemented feature to stack popups within the browser
(required for modal popups like having a formbox inside a popup)
* Debug; removed some verbose outputs
* Page popup widget; now supports triggers through javascript (enabling
buttons to open popups defined in the page popup)
* Article link options types; by default the media item selector is
enabled, but can now also be turned off if needed.
* Formbox; improved UX; shows a wait while loading image while the
server is processing the request, and makes the button's opacity 0.5
while processing.
* Formbox; dial box that shows the validation errors is now renders in a
"pushed" popup stack. When clicking on the dismiss button, it will pop
the stack back to the previous popup instead of tearing down all popups