Repeater (List)
* Repeater (List); the repeater (list) widget now stores lookup values
of each item in the "global" dynamic lookups variable, and removes that
variable when its finished processing each item. Practical to allow
shortcodes inside an item to use the lookup entries (widget, as well as
item).
* Repeater (List); the repeater (list) widget now applies a filter that
allows plugins to tune the result of the widget before handling it over
to the framework (used in the adwords plugin p001 which stores the
output in a CSV file instead of showing it on the screen)