DataProtection
* DataProtection; cookie wall is now implemented and works (screen does
not yet look pretty, but it functions like it should)
* DataProtection; added the logic to recurvisvely build a "tree" of
protected data within a system
* DataProtection; added generic function that allows component to ask
for permission to execute a particular action based. The data protection
service is responsible for figuring out the answer. Example is that the
framework could wonder if its allowed to return cached data for a user
(answer is no, if the consent was not given, if the cookie wall is set
as the enforced data protection type).