Data Protection
* Data Protection; cookiewall activity is now a semantic unit; if the
cookiewal is on, this will control whether other activities will be on
or not (rather than fine grained component control).
* Data Protection; the url of the privacy policy is now centralized
* Data Protection; controller now controls whether activities are
enabled or disabled (including the cookie wall itself). There's no more
fine grained control of how the user consent is configurable; its either
always on, or its determined by the consent of the cookie wall activity.