Cleanup, Data Protection, Cookie wall, Privacy Policy (WP 4.9.6), PrivacyConsentFormItem, Popups, GotoSheet, FormBox
* Cleanup; removed obsolete references to old servers
* Cleanup; removed obsolete seotab code
* Data protection; upgraded various http references to https
* Cookie wall; fixed nxs_dataprotection_getcookiewallconsenttext
* Privacy Policy (WP 4.9.6); rewrote parts of the code that are now
taken care of by the latest WP version 4.9.6 (improving compatibility)
* Data Protection; split code and popups from one junk to a Cookie Wall
piece and a Component Control piece
* PrivacyConsentFormItem; introduced new widget PrivacyConsentFormItem
which can be used in forms; the form won't submit data to the server if
this widget is used, unless the user has explicitly gave a consent this
is ok. Hooks to the new privacy policy feature of WP (4.9.6)
* Popups; added generic helper factory function to quickly generate
popups with only a notification;
nxs_popup_factory_createnotificationoptions. Very practical for quickly
rendering alternative flows if for example certain criteria/dependencies
are not there
* GotoSheet; the gotosheet popup optiontype now also supports a custom
piece of html which is rendered to the left side of the button that goes
to the new sheet
* Shortcodes; added "privacypolicy_title, privacypolicy_text and
privacypolicy_url operators for the nxs_string shortcode
* ContactItemFileAttachment; is now again a component that can be turned
on/off by the controller
* FormBox; now implements a way to check for explicit consents prior to
sending
* FormBox; now pre-loads the formbox.js script after 3 seconds; won't
slow down the page loading, and will be faster for people submitting
forms
* FormItemCaptcha; is now a Component that can be turned on/off by the
controller
* FormItemCheckbox; reverted the label from tinymce rich text to regular
flat text input (one line). The tinymce editor didn't fit the needs to
have a link and ask for consent (one reason was the editor was way too
clunky and very impractical for like 99% of the cases where people only
wanted to have a simply checkbox and no consent, and the most important
reason is that the form validation occurs server side; prior to the
given consent the fields would have reached the server which would be a
violation of the GDPR rule)
* Cleanup; removed various conceptual widgets (BusLogo, CommercialMsgs,
SocialAccounts, TaxPageSlider)