Page Slider
* Page Slider; fixed hiding the static positioned footer bar; in the
scenario where a page decorator contained both a page slider and page
background (further down in the stack) then the static positioned footer
bar would still show even if you configured it to be hidden. Was caused
by the page background overriding a global variable. Note that combining
a page slider and background image is kind of weird but possible
(silvio)