shortcodes urlencode, rawurlencode
* shortcodes; added rawurlencode operator of nxs_string shortcode
* shortcodes; the urlencode operator of nxs_string shortcode now invokes
the rawurlencode function instead of the urlencode one. This new
implementation helps prevents useless 301's from taking place if you use
the output in URLs that have "spaces" in them; in the old implementation
those spaces would then have been encoded to "+" (plus signs); those +
signs would result in the browser 301-ing the urls to the same url with
the + being replaced with %20's. Saves another useless roundtrip,
improving the UX and lowering the stress on the webserver.