Shortcodes
* Shortcodes; added the "ishandheld", "!ishandheld" and "notishandheld"
operator in the nxs_bool shortcode.
* Shortcodes; the file_get_contents and filegetcontents operators of the
nxs_string shortcode will only work if a filter
(nxs_string_file_get_contents_enabled) has provided explicit permission
to use the shortcode. Default is that this functionality is blocked.
* Shortcodes; the httpok operator of the nxs_bool shortcode will only
work if a filter (nxs_bool_httpok_enabled) has provided explicit
permission to use the shortcode. Default is that this functionality is
blocked.