Twitter, Bootstrap, Repeater, Embed, EmbedRepeater, Social sharing
* Twitter; the twitter widget no longer worked because of changes made
by Twitter. We now reference the support in the twitter widget is you
want to render tweets instead.
* Bootstrap; not sure why the bootstrap framework was embedded in the
nexus framework, but it was never referenced, cleanup.
* The widgets Section, Social sharing, Signpost, Tumbler, Radial,
Target, Repeater, Embed, EmbedRepeater are no longer enabled by default
for users as they rise more questions and doesn't solve anything; we aim
for a clean "do make me think" framework. Widgets invented in the past
by ourselves that don't make sense for new users should not be enabled
to avoid confusion. They are still in the framework to ensure backwards
compatibility
*