Extensions, My account
* Extensions; the extensions plug showing at the top when logged into
the WP environment now links to the "dashboardextensions" popup. It
renders a link helping people to navigate to the WP plugin repository of
WP. It no longer shows the plugins offered on our hosting environment.
* My accounts; added a link to the upcoming "my account" of nexusthemes
where people can download updates, read news, and manage their account
and interact with the community of other users. The "my accounts" link
has a filter, allowing people to disable the link when needed.