Shortcodes
* Shortcodes; shortcodes can now conditionally be evaluated depending on
whether or not the runtime is, or is not invoked on a particular
hostname. When the shortcode is not evaluated it returns the requested
shortcode in "raw" form.  For now its implemented for nxs_string
shortcodes only, when practical can be implemented for other shortcodes
too.
* Shortcodes; the "guid" nxs_string operator now also supports returning
one unique guid per singleton within a given runtime, if a singleton
identification is specified. If the singleton identification is not
specified, will return a unique guid for each invocation similar to how
the operator used to work.